+++ src/test/scala/zio/kafka/AdminSpec.scala
@@ -1,7 +1,7 @@
 package zio.kafka.admin
 
 import org.apache.kafka.clients.consumer.ConsumerRecord
-import zio.Chunk
+import zio.{ Chunk, Has }
 import zio.blocking.Blocking
 import zio.clock.Clock
 import zio.kafka.KafkaTestUtils
@@ -144,7 +144,7 @@ object AdminSpec extends DefaultRunnableSpec {
 
           for {
             _       <- client.createTopics(List(AdminClient.NewTopic("topic8", 3, 1)))
-            _       <- produceMany(topic, kvs).provideSomeLayer[Kafka with Blocking with Clock](stringProducer)
+            _       <- produceMany(topic, kvs).provideSomeLayer[Has[Kafka] with Blocking with Clock](producer)
             offsets <- client.listOffsets(
                          (0 until 3).map(i => TopicPartition(topic, i) -> OffsetSpec.LatestSpec).toMap
                        )
@@ -165,7 +165,7 @@ object AdminSpec extends DefaultRunnableSpec {
           def consumeAndCommit(count: Long) =
             Consumer
               .subscribeAnd(Subscription.Topics(Set(topic)))
-              .partitionedStream[Kafka with Blocking with Clock, String, String](Serde.string, Serde.string)
+              .partitionedStream[Has[Kafka] with Blocking with Clock, String, String](Serde.string, Serde.string)
               .flatMapPar(partitionCount)(_._2)
               .take(count)
               .transduce(ZTransducer.collectAllN(Int.MaxValue))
@@ -177,7 +177,7 @@ object AdminSpec extends DefaultRunnableSpec {
                 offsetBatch.commit.as(records)
               }
               .runCollect
-              .provideSomeLayer[Kafka with Blocking with Clock](consumer(consumerGroupID, "topic9"))
+              .provideSomeLayer[Has[Kafka] with Blocking with Clock](consumer(consumerGroupID, "topic9"))
 
           def toMap(records: Chunk[ConsumerRecord[String, String]]): Map[Int, List[(Long, String, String)]] =
             records.toList
@@ -187,7 +187,7 @@ object AdminSpec extends DefaultRunnableSpec {
 
           for {
             _                    <- client.createTopics(List(AdminClient.NewTopic(topic, partitionCount, 1)))
-            _                    <- produceMany(topic, kvs).provideSomeLayer[Kafka with Blocking with Clock](stringProducer)
+            _                    <- produceMany(topic, kvs).provideSomeLayer[Has[Kafka] with Blocking with Clock](producer)
             records              <- consumeAndCommit(msgCount.toLong).map(toMap)
             endOffsets           <- client.listOffsets((0 until partitionCount).map(i => p(i) -> OffsetSpec.LatestSpec).toMap)
             _                    <- client.alterConsumerGroupOffsets(
@@ -211,10 +211,10 @@ object AdminSpec extends DefaultRunnableSpec {
         def consumeAndCommit(count: Long, topic: String, groupId: String) =
           Consumer
             .subscribeAnd(Subscription.Topics(Set(topic)))
-            .plainStream[Kafka with Blocking with Clock, String, String](Serde.string, Serde.string)
+            .plainStream[Has[Kafka] with Blocking with Clock, String, String](Serde.string, Serde.string)
             .take(count)
             .foreach(_.offset.commit)
-            .provideSomeLayer[Kafka with Blocking with Clock](consumer(groupId, topic))
+            .provideSomeLayer[Has[Kafka] with Blocking with Clock](consumer(groupId, topic))
 
         KafkaTestUtils.withAdmin { client =>
           for {
@@ -226,7 +226,7 @@ object AdminSpec extends DefaultRunnableSpec {
             msgConsume             = 15
             kvs                    = (1 to msgCount).toList.map(i => (s"key$i", s"msg$i"))
             _                     <- client.createTopics(List(AdminClient.NewTopic(topic, 1, 1)))
-            _                     <- produceMany(topic, kvs).provideSomeLayer[Kafka with Blocking](stringProducer)
+            _                     <- produceMany(topic, kvs).provideSomeLayer[Has[Kafka] with Blocking](producer)
             _                     <- consumeAndCommit(msgConsume.toLong, topic, groupId)
             offsets               <- client.listConsumerGroupOffsets(
                                        groupId,
