+++ src/test/scala/zio/kafka/KafkaTestUtils.scala
@@ -17,41 +17,48 @@ import zio.kafka.serde.{ Deserializer, Serde, Serializer }
 import zio.kafka.producer._
 
 object KafkaTestUtils {
-  type StringProducer = Producer[Any, String, String]
 
-  val producerSettings: ZIO[Kafka, Nothing, ProducerSettings] =
-    ZIO.access[Kafka](_.get[Kafka.Service].bootstrapServers).map(ProducerSettings(_))
+  val producerSettings: ZIO[Has[Kafka], Nothing, ProducerSettings] =
+    ZIO.access[Has[Kafka]](_.get[Kafka].bootstrapServers).map(ProducerSettings(_))
 
-  val stringProducer: ZLayer[Kafka with Blocking, Throwable, StringProducer] =
+  val producer: ZLayer[Has[Kafka] with Blocking, Throwable, Has[Producer]] =
     (ZLayer.fromEffect(producerSettings) ++ ZLayer.succeed(Serde.string: Serializer[Any, String])) ++ ZLayer
       .identity[Blocking] >>>
-      Producer.live[Any, String, String]
+      Producer.live
 
   def produceOne(
     topic: String,
     key: String,
     message: String
-  ): ZIO[Blocking with StringProducer, Throwable, RecordMetadata] =
-    Producer.produce[Any, String, String](new ProducerRecord(topic, key, message))
+  ): ZIO[Blocking with Has[Producer], Throwable, RecordMetadata] =
+    Producer.produce[Any, String, String](new ProducerRecord(topic, key, message), Serde.string, Serde.string)
 
   def produceMany(
     topic: String,
     partition: Int,
     kvs: Iterable[(String, String)]
-  ): ZIO[Blocking with StringProducer, Throwable, Chunk[RecordMetadata]] =
+  ): ZIO[Blocking with Has[Producer], Throwable, Chunk[RecordMetadata]] =
     Producer
-      .produceChunk[Any, String, String](Chunk.fromIterable(kvs.map { case (k, v) =>
-        new ProducerRecord(topic, partition, null, k, v)
-      }))
+      .produceChunk[Any, String, String](
+        Chunk.fromIterable(kvs.map { case (k, v) =>
+          new ProducerRecord(topic, partition, null, k, v)
+        }),
+        Serde.string,
+        Serde.string
+      )
 
   def produceMany(
     topic: String,
     kvs: Iterable[(String, String)]
-  ): ZIO[Blocking with StringProducer, Throwable, Chunk[RecordMetadata]] =
+  ): ZIO[Blocking with Has[Producer], Throwable, Chunk[RecordMetadata]] =
     Producer
-      .produceChunk[Any, String, String](Chunk.fromIterable(kvs.map { case (k, v) =>
-        new ProducerRecord(topic, k, v)
-      }))
+      .produceChunk[Any, String, String](
+        Chunk.fromIterable(kvs.map { case (k, v) =>
+          new ProducerRecord(topic, k, v)
+        }),
+        Serde.string,
+        Serde.string
+      )
 
   def consumerSettings(
     groupId: String,
@@ -59,7 +66,7 @@ object KafkaTestUtils {
     allowAutoCreateTopics: Boolean = true,
     offsetRetrieval: OffsetRetrieval = OffsetRetrieval.Auto()
   ) =
-    ZIO.access[Kafka] { kafka: Kafka =>
+    ZIO.access[Has[Kafka]] { kafka: Has[Kafka] =>
       ConsumerSettings(kafka.get.bootstrapServers)
         .withGroupId(groupId)
         .withClientId(clientId)
@@ -82,14 +89,14 @@ object KafkaTestUtils {
     offsetRetrieval: OffsetRetrieval = OffsetRetrieval.Auto(),
     allowAutoCreateTopics: Boolean = true,
     diagnostics: Diagnostics = Diagnostics.NoOp
-  ): ZLayer[Kafka with Clock with Blocking, Throwable, Consumer] =
+  ): ZLayer[Has[Kafka] with Clock with Blocking, Throwable, Has[Consumer]] =
     (consumerSettings(groupId, clientId, allowAutoCreateTopics, offsetRetrieval).toLayer ++
       ZLayer.requires[Clock with Blocking] ++
       ZLayer.succeed(diagnostics)) >>> Consumer.live
 
   def consumeWithStrings[RC](groupId: String, clientId: String, subscription: Subscription)(
     r: (String, String) => URIO[RC, Unit]
-  ): RIO[RC with Blocking with Clock with Kafka, Unit] =
+  ): RIO[RC with Has[Kafka] with Blocking with Clock, Unit] =
     consumerSettings(groupId, clientId).flatMap { settings =>
       Consumer.consumeWith(
         settings,
@@ -99,16 +106,16 @@ object KafkaTestUtils {
       )(r)
     }
 
-  def adminSettings: ZIO[Kafka, Nothing, AdminClientSettings] =
-    ZIO.access[Kafka](_.get[Kafka.Service].bootstrapServers).map(AdminClientSettings(_))
+  def adminSettings: ZIO[Has[Kafka], Nothing, AdminClientSettings] =
+    ZIO.access[Has[Kafka]](_.get[Kafka].bootstrapServers).map(AdminClientSettings(_))
 
-  def withAdmin[T](f: AdminClient => RIO[Clock with Kafka with Blocking, T]) =
+  def withAdmin[T](f: AdminClient => RIO[Clock with Has[Kafka] with Blocking, T]) =
     for {
       settings <- adminSettings
       fRes     <- AdminClient
                     .make(settings)
                     .use(client => f(client))
-                    .provideSomeLayer[Kafka](Clock.live ++ Blocking.live)
+                    .provideSomeLayer[Has[Kafka]](Clock.live ++ Blocking.live)
     } yield fRes
 
   def randomThing(prefix: String): Task[String] =
