+++ src/test/scala/zio/kafka/embedded/Kafka.scala
@@ -1,0 +1,31 @@
+package zio.kafka.embedded
+
+import io.github.embeddedkafka.{ EmbeddedK, EmbeddedKafka, EmbeddedKafkaConfig }
+import zio._
+
+trait Kafka {
+  def bootstrapServers: List[String]
+  def stop(): UIO[Unit]
+}
+
+object Kafka {
+
+  case class EmbeddedKafkaService(embeddedK: EmbeddedK) extends Kafka {
+    override def bootstrapServers: List[String] = List(s"localhost:${embeddedK.config.kafkaPort}")
+    override def stop(): UIO[Unit]              = ZIO.effectTotal(embeddedK.stop(true))
+  }
+
+  case object DefaultLocal extends Kafka {
+    override def bootstrapServers: List[String] = List(s"localhost:9092")
+    override def stop(): UIO[Unit]              = UIO.unit
+  }
+
+  val embedded: ZLayer[Any, Throwable, Has[Kafka]] = ZLayer.fromManaged {
+    implicit val embeddedKafkaConfig = EmbeddedKafkaConfig(
+      customBrokerProperties = Map("group.min.session.timeout.ms" -> "500", "group.initial.rebalance.delay.ms" -> "0")
+    )
+    ZManaged.make(ZIO.effect(EmbeddedKafkaService(EmbeddedKafka.start())))(_.stop())
+  }
+
+  val local: ZLayer[Any, Nothing, Has[Kafka]] = ZLayer.succeed(DefaultLocal)
+}
